[core] Modularise `uint256`
This patch modularises the `uint256` implementation such that we can
more easily support it on multiple platforms. The library remains
header-only. The implementation is split into the following components:

* The core implementation (and user-facing interface) remains in
  `uint256.hpp`
* Platform-specific intrinsics are put into `uint256/intrinsics/`, and
  CMake has been set up to select the appropriate header for the target
  platform. The header `uint256/intrinsics.hpp` defines the signatures
  of all the intrinsics that each backend implementation must define.
* The constexpr-versions of the binary operations have been hoisted to
  `uint256/binary_operations.hpp`.
* `uint256/types.hpp` contains the shared types used by the core
  implementation, the constexpr binary operations, and the intrinsics.

The special intrinsics implementation `uint256/intrinsics/portable.hpp`
defines the intrinsics interface in terms of the `constexpr` binary
operations.
