+++ category/core/runtime/uint256.hpp
@@ -18,6 +18,9 @@
 #include <category/core/assert.h>
 #include <category/core/likely.h>
 #include <category/core/runtime/uint128.hpp>
+#include <category/core/runtime/uint256/intrinsics.hpp>
+#include <category/core/runtime/uint256/portable.hpp>
+#include <category/core/runtime/uint256/types.hpp>
 
 #include <algorithm>
 #include <array>
@@ -27,19 +30,10 @@
 #include <cstdint>
 #include <cstring>
 #include <format>
-#include <immintrin.h>
 #include <limits>
 #include <stdexcept>
 #include <string>
 #include <type_traits>
-
-#ifndef __AVX2__
-    #error "Target architecture must support AVX2"
-#endif
-
-#ifndef __BMI2__
-    #error "Target architecture must support BMI2 (for MULX)"
-#endif
 
 // GCC's overeager SLP vectorizer sometimes pessimizes code. For functions that
 // are particularly sensitive about this (such as multiplication), the
@@ -56,54 +50,38 @@
 
 namespace monad
 {
-    struct uint256_t;
-
-    [[gnu::always_inline]] constexpr inline uint64_t
-    force(uint64_t expr) noexcept
+    // The constructor `explicit(true) uint256_t(__m256i x)` further below is
+    // only meaningful on AVX2 targets. We use a template-trick to disable it on
+    // non-AVX2 targets.
+#if defined(__AVX2__)
+    // GCC complains about ignored attributes on `__m256i` when used
+    // directly as a parameter type. We can fix this using another layer of
+    // indirection by wrapping it in a trivial struct that can only be
+    // constructed from a `__m256i` value. This is a compile-time artifact,
+    // meaning it is never materialized at runtime.
+    struct m256i_wrapper
     {
-        if !consteval {
-            asm("" : "+r"(expr));
-        }
-        return expr;
-    }
+        __m256i const v;
 
-    template <typename T>
-    struct result_with_carry
-    {
-        T value;
-        bool carry;
+        m256i_wrapper() = delete;
 
-        // Only used in unit tests
-        friend inline bool operator==(
-            result_with_carry const &lhs, result_with_carry const &rhs) noexcept
-            requires std::equality_comparable<T>
+        explicit(false) constexpr m256i_wrapper(__m256i x)
+            : v(x)
         {
-            return lhs.value == rhs.value && lhs.carry == rhs.carry;
         }
     };
 
-    [[gnu::always_inline]]
-    constexpr inline result_with_carry<uint64_t> addc_constexpr(
-        uint64_t const lhs, uint64_t const rhs, bool const carry_in) noexcept
-    {
-        uint64_t const sum = lhs + rhs;
-        bool carry_out = sum < lhs;
-        uint64_t const sum_carry = sum + carry_in;
-        carry_out |= sum_carry < sum;
-        return result_with_carry{.value = sum_carry, .carry = carry_out};
-    }
+    template <typename T>
+    concept m256i_native = std::convertible_to<T, m256i_wrapper>;
+    using m256i = __m256i;
+#else
+    // On non-AVX2 targets we interpret `__m256i` as four words.
+    template <typename T>
+    concept m256i_native = false;
+    using m256i = words_t<4>;
+#endif
 
-    [[gnu::always_inline]]
-    inline result_with_carry<uint64_t> addc_intrinsic(
-        uint64_t const lhs, uint64_t const rhs, bool const carry_in) noexcept
-    {
-        static_assert(sizeof(unsigned long long) == sizeof(uint64_t));
-        unsigned long long carry_out = 0;
-        uint64_t const sum_carry =
-            __builtin_addcll(lhs, rhs, carry_in, &carry_out);
-        return result_with_carry{
-            .value = sum_carry, .carry = static_cast<bool>(carry_out)};
-    }
+    struct uint256_t;
 
     [[gnu::always_inline]]
     constexpr inline result_with_carry<uint64_t>
@@ -118,30 +96,6 @@ namespace monad
     }
 
     [[gnu::always_inline]] constexpr inline result_with_carry<uint64_t>
-    subb_constexpr(
-        uint64_t const lhs, uint64_t const rhs, bool const borrow_in) noexcept
-    {
-        uint64_t const sub = lhs - rhs;
-        bool borrow_out = rhs > lhs;
-        uint64_t const sub_borrow = sub - borrow_in;
-        borrow_out |= borrow_in > sub;
-        return result_with_carry{.value = sub_borrow, .carry = borrow_out};
-    }
-
-    [[gnu::always_inline]] inline result_with_carry<uint64_t> subb_intrinsic(
-        uint64_t const lhs, uint64_t const rhs, bool const borrow_in) noexcept
-    {
-        static_assert(sizeof(unsigned long long) == sizeof(uint64_t));
-        unsigned long long borrow_out = 0;
-        uint64_t const sub_borrow =
-            __builtin_subcll(lhs, rhs, borrow_in, &borrow_out);
-        // If we do not force the result here, clang replaces the sub/sbb chain
-        // with a long series of comparisons and flag logic which is worse
-        return result_with_carry{
-            .value = force(sub_borrow), .carry = static_cast<bool>(borrow_out)};
-    }
-
-    [[gnu::always_inline]] constexpr inline result_with_carry<uint64_t>
     subb(uint64_t const lhs, uint64_t const rhs, bool const borrow_in) noexcept
     {
         if consteval {
@@ -150,24 +104,6 @@ namespace monad
         else {
             return subb_intrinsic(lhs, rhs, borrow_in);
         }
-    }
-
-    [[gnu::always_inline]]
-    inline uint64_t shld_intrinsic(
-        uint64_t high, uint64_t const low, uint8_t const shift) noexcept
-    {
-        asm("shldq %[shift], %[low], %[high]"
-            : [high] "+r"(high)
-            : [low] "r"(low), [shift] "c"(shift)
-            : "cc");
-        return high;
-    }
-
-    [[gnu::always_inline]]
-    inline constexpr uint64_t shld_constexpr(
-        uint64_t const high, uint64_t const low, uint8_t const shift) noexcept
-    {
-        return (high << shift) | ((low >> 1) >> (63 - shift));
     }
 
     [[gnu::always_inline]]
@@ -183,24 +119,6 @@ namespace monad
     }
 
     [[gnu::always_inline]]
-    inline uint64_t shrd_intrinsic(
-        uint64_t const high, uint64_t low, uint8_t const shift) noexcept
-    {
-        asm("shrdq %[shift], %[high], %[low]"
-            : [low] "+r"(low)
-            : [high] "r"(high), [shift] "c"(shift)
-            : "cc");
-        return low;
-    }
-
-    [[gnu::always_inline]]
-    inline constexpr uint64_t shrd_constexpr(
-        uint64_t const high, uint64_t const low, uint8_t const shift) noexcept
-    {
-        return (low >> shift) | ((high << 1) << (63 - shift));
-    }
-
-    [[gnu::always_inline]]
     inline constexpr uint64_t
     shrd(uint64_t const high, uint64_t const low, uint8_t const shift) noexcept
     {
@@ -210,40 +128,6 @@ namespace monad
         else {
             return shrd_intrinsic(high, low, shift);
         }
-    }
-
-    template <typename Q, typename R = Q>
-    struct div_result
-    {
-        Q quot;
-        R rem;
-
-        // Only used in unit tests
-        friend inline bool
-        operator==(div_result const &lhs, div_result const &rhs) noexcept
-            requires std::equality_comparable<Q> && std::equality_comparable<R>
-        {
-            return lhs.quot == rhs.quot && lhs.rem == rhs.rem;
-        }
-    };
-
-    [[gnu::always_inline]]
-    constexpr inline div_result<uint64_t>
-    div_constexpr(uint64_t u_hi, uint64_t u_lo, uint64_t const v) noexcept
-    {
-        using u128 = unsigned __int128;
-        auto const u = static_cast<u128>(uint128_t{u_lo, u_hi});
-        auto const quot = static_cast<uint64_t>(u / v);
-        auto const rem = static_cast<uint64_t>(u % v);
-        return {.quot = quot, .rem = rem};
-    }
-
-    [[gnu::always_inline]]
-    inline div_result<uint64_t>
-    div_intrinsic(uint64_t u_hi, uint64_t u_lo, uint64_t const v) noexcept
-    {
-        asm("div %[v]" : "+d"(u_hi), "+a"(u_lo) : [v] "r"(v));
-        return {.quot = u_lo, .rem = u_hi};
     }
 
     [[gnu::always_inline]]
@@ -299,8 +183,9 @@ namespace monad
         {
         }
 
+        template <m256i_native M>
         [[gnu::always_inline]]
-        explicit(true) uint256_t(__m256i x) noexcept
+        explicit(true) uint256_t(M x) noexcept
         {
             // Clang sometimes miscompiles the std::bit_cast equivalent into a
             // much slower version, so we prefer memcpy here
@@ -308,9 +193,9 @@ namespace monad
         }
 
         [[gnu::always_inline]]
-        inline __m256i to_avx() const noexcept
+        inline m256i to_avx() const noexcept
         {
-            __m256i result;
+            m256i result;
             std::memcpy(&result, &words_, sizeof(result));
             return result;
         }
@@ -785,31 +670,6 @@ namespace monad
     }
 
     [[gnu::always_inline]]
-    inline constexpr void mulx_constexpr(
-        uint64_t const x, uint64_t const y, uint64_t &r_hi,
-        uint64_t &r_lo) noexcept
-    {
-        using u128 = unsigned __int128;
-        u128 const prod = static_cast<u128>(x) * static_cast<u128>(y);
-        r_hi = static_cast<uint64_t>(prod >> u128{64});
-        r_lo = static_cast<uint64_t>(prod);
-    }
-
-    [[gnu::always_inline]]
-    inline void mulx_intrinsic(
-        uint64_t const x, uint64_t const y, uint64_t &r_hi,
-        uint64_t &r_lo) noexcept
-    {
-        /*
-        uint64_t hi;
-        uint64_t lo;
-        */
-        asm("mulx %[x], %[lo], %[hi]"
-            : [hi] "=r"(r_hi), [lo] "=r"(r_lo)
-            : [x] "r"(x), [y] "d"(y));
-    }
-
-    [[gnu::always_inline]]
     inline constexpr void mulx(
         uint64_t const x, uint64_t const y, uint64_t &r_hi,
         uint64_t &r_lo) noexcept
@@ -820,243 +680,6 @@ namespace monad
         else {
             return mulx_intrinsic(x, y, r_hi, r_lo);
         }
-    }
-
-    template <size_t M>
-    using words_t = std::array<uint64_t, M>;
-
-    [[gnu::always_inline]]
-    inline void adc_3(
-        uint64_t x_2, uint64_t x_1, uint64_t x_0, uint64_t const y_1,
-        uint64_t const y_0, uint64_t &r_2, uint64_t &r_1,
-        uint64_t &r_0) noexcept
-    {
-        asm("addq %[y_0], %[x_0]\n"
-            "adcq %[y_1], %[x_1]\n"
-            "adcq $0, %[x_2]"
-            : [x_0] "+r"(x_0), [x_1] "+r"(x_1), [x_2] "+r"(x_2)
-            : [y_0] "r"(y_0), [y_1] "r"(y_1)
-            : "cc");
-        r_2 = x_2;
-        r_1 = x_1;
-        r_0 = x_0;
-    }
-
-    [[gnu::always_inline]]
-    inline void adc_2(
-        uint64_t x_1, uint64_t x_0, uint64_t const y_0, uint64_t &r_1,
-        uint64_t &r_0) noexcept
-    {
-        asm("addq %[y_0], %[x_0]\n"
-            "adcq $0, %[x_1]"
-            : [x_0] "+r"(x_0), [x_1] "+r"(x_1)
-            : [y_0] "r"(y_0)
-            : "cc");
-        r_1 = x_1;
-        r_0 = x_0;
-    }
-
-    [[gnu::always_inline]]
-    inline void adc_2(
-        uint64_t x_1, uint64_t x_0, uint64_t const y_1, uint64_t const y_0,
-        uint64_t &r_1, uint64_t &r_0) noexcept
-    {
-        asm("addq %[y_0], %[x_0]\n"
-            "adcq %[y_1], %[x_1]"
-            : [x_0] "+r"(x_0), [x_1] "+r"(x_1)
-            : [y_0] "r"(y_0), [y_1] "r"(y_1)
-            : "cc");
-        r_1 = x_1;
-        r_0 = x_0;
-    }
-
-    template <size_t I, size_t R, size_t M>
-    [[gnu::always_inline]]
-    inline void mul_line_recur(
-        words_t<M> const &x, uint64_t const y, words_t<R> &__restrict__ result,
-        uint64_t carry) noexcept
-    {
-        if constexpr (I < std::min(R, M)) {
-            if constexpr (I + 1 < R) {
-                uint64_t hi;
-                uint64_t lo;
-                mulx(x[I], y, hi, lo);
-                adc_2(
-                    // Input 1
-                    hi,
-                    lo,
-                    // Input 2
-                    carry,
-                    // Output
-                    carry,
-                    result[I]);
-                mul_line_recur<I + 1, R, M>(x, y, result, carry);
-            }
-            else {
-                result[I] = y * x[I] + carry;
-            }
-        }
-        else if constexpr (M < R) {
-            result[M] = carry;
-        }
-    }
-
-    // result[0 .. min(M + 1, R)) = y * x[0 .. M)
-    template <size_t R, size_t M>
-    [[gnu::always_inline]]
-    inline void mul_line(
-        words_t<M> const &x, uint64_t const y,
-        words_t<R> &__restrict__ result) noexcept
-    {
-        uint64_t carry;
-        mulx(y, x[0], carry, result[0]);
-
-        mul_line_recur<1, R, M>(x, y, result, carry);
-    }
-
-    template <size_t J, size_t I, size_t R, size_t M>
-    [[gnu::always_inline]]
-    inline void mul_add_line_recur(
-        words_t<M> const &x, uint64_t const y_i,
-        words_t<R> &__restrict__ result, uint64_t c_hi, uint64_t c_lo) noexcept
-    {
-        if constexpr (J + 1 < M && I + J < R) {
-            if constexpr (I + J + 2 < R) {
-                // We need c_lo, c_hi
-                uint64_t hi;
-                uint64_t lo;
-                mulx(x[J + 1], y_i, hi, lo);
-                adc_3(
-                    // Input 1
-                    hi,
-                    lo,
-                    result[I + J],
-                    // Input 2
-                    c_hi,
-                    c_lo,
-                    // Result
-                    c_hi,
-                    c_lo,
-                    result[I + J]);
-            }
-            else if constexpr (I + J + 1 < R) {
-                // We only need c_lo
-                uint64_t const lo = x[J + 1] * y_i;
-                adc_2(
-                    // Input 1
-                    lo,
-                    result[I + J],
-                    // Input 2
-                    c_hi,
-                    c_lo,
-                    // Output
-                    c_lo,
-                    result[I + J]);
-            }
-            else {
-                // We're done, we don't need subsequent results
-                result[I + J] += c_lo;
-            }
-            mul_add_line_recur<J + 1, I, R, M>(x, y_i, result, c_hi, c_lo);
-        }
-        else {
-            if constexpr (I + M < R) {
-
-                adc_2(
-                    // Input 1
-                    c_hi,
-                    c_lo,
-                    // Input 2
-                    result[I + M - 1],
-                    // Output
-                    result[I + M],
-                    result[I + M - 1]);
-            }
-            else if constexpr (I + M < R + 1) {
-                result[I + M - 1] += c_lo;
-            }
-        }
-    }
-
-    // result[i .. min(i + M + 1, R)) += y_i * x[0 .. M)
-    template <size_t I, size_t R, size_t M>
-    [[gnu::always_inline]]
-    inline void mul_add_line(
-        words_t<M> const &x, uint64_t const y_i,
-        words_t<R> &__restrict__ result) noexcept
-    {
-        // A naive implementation would use a single carry variable. However
-        // this means on every iteration we compute
-        //     result[i+j] = result[i+j] + prod_lo + carry
-        // which needs to propagate two carry bits. This causes a slew of
-        // setb/xor/movxz instructions to be inserted.
-        // Instead, we widen the carry and skew the loop so that every iteration
-        // computes
-        //     result[i+j] = result[i+j] + c_lo
-        //     c_lo = prod_lo + c_hi (+ carry)
-        //     c_hi = prod_hi (+ carry)
-        uint64_t c_hi;
-        uint64_t c_lo;
-
-        if constexpr (I + 1 < R) {
-            mulx(x[0], y_i, c_hi, c_lo);
-        }
-        else {
-            c_hi = 0;
-            c_lo = x[0] * y_i;
-        }
-        mul_add_line_recur<0, I, R, M>(x, y_i, result, c_hi, c_lo);
-    }
-
-    template <size_t I, size_t R, size_t M, size_t N>
-    [[gnu::always_inline]]
-    inline void truncating_mul_runtime_recur(
-        words_t<M> const &x, words_t<N> const &y,
-        words_t<R> &__restrict__ result) noexcept
-    {
-        if constexpr (I < N) {
-            mul_add_line<I, R, M>(x, y[I], result);
-            truncating_mul_runtime_recur<I + 1, R, M>(x, y, result);
-        }
-    }
-
-    template <size_t R, size_t M, size_t N>
-    [[gnu::always_inline]]
-    inline words_t<R>
-    truncating_mul_runtime(words_t<M> const &x, words_t<N> const &y) noexcept
-        requires(0 < R && 0 < M && 0 < N && R <= M + N)
-    {
-        words_t<R> result;
-        mul_line<R>(x, y[0], result);
-        truncating_mul_runtime_recur<1, R, M, N>(x, y, result);
-        return result;
-    }
-
-    template <size_t R, size_t M, size_t N>
-    [[gnu::always_inline]]
-    inline constexpr words_t<R>
-    truncating_mul_constexpr(words_t<M> const &x, words_t<N> const &y) noexcept
-        requires(0 < R && R <= M + N)
-    {
-        words_t<R> result{0};
-        for (size_t j = 0; j < N; j++) {
-            uint64_t carry = 0;
-            for (size_t i = 0; i < M && i + j < R; i++) {
-                uint64_t hi;
-                uint64_t lo;
-                mulx(x[i], y[j], hi, lo);
-
-                auto const [s0, c0] = addc(lo, result[i + j], false);
-                auto const [s1, c1] = addc(s0, carry, false);
-                result[i + j] = s1;
-                auto const [s2, c2] = addc(hi, c0, c1);
-                carry = s2;
-            }
-            if (j + M < R) {
-                result[j + M] = carry;
-            }
-        }
-        return result;
     }
 
     /**
@@ -1719,3 +1342,5 @@ struct std::formatter<monad::uint256_t>
         return std::format_to(ctx.out(), "0x{}", v.to_string(16));
     }
 };
+
+#undef MONAD_NO_VECTORIZE
